backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
marshmallow==3.20.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Flask-Limiter==3.5.0
pytest==7.4.3
pytest-flask==1.3.0
